pyproject.toml
[tool.poetry]
name = "cue-backend"
version = "0.1.0"
description = "AWS code and infrastucture for the CUE enterprise product"
authors = ["Davis Gaunt <bdg0015@uah.edu>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "~0.112.2"
pydantic = "~2.8.2"
uvicorn = "~0.30.6"
python-dotenv = "~1.0.1"
mangum = "~0.17.0"
boto3 = "~1.35.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
